add Exception Handler add special exception for balance

+++ src/main/java/ru/kolesnikov/storemanagement/controller/dto/receiptdetails/ReceiptDetailDTORequest.java
@@ -2,8 +2,9 @@ package ru.kolesnikov.storemanagement.controller.dto.receiptdetails;
 
 import javax.validation.constraints.NotBlank;
 import java.math.BigDecimal;
+import java.math.BigInteger;
 
-public record ReceiptDetailDTORequest(@NotBlank BigDecimal quantity,
+public record ReceiptDetailDTORequest(@NotBlank BigInteger quantity,
                                       @NotBlank BigDecimal price,
                                       @NotBlank String barcode) {
 }
