+++ src/main/java/ru/kolesnikov/storemanagement/controller/dto/receiptdetails/ReceiptDetailDTOResponse.java
@@ -1,9 +1,10 @@
 package ru.kolesnikov.storemanagement.controller.dto.receiptdetails;
 
 import java.math.BigDecimal;
+import java.math.BigInteger;
 
 public record ReceiptDetailDTOResponse(Long id,
-                                       BigDecimal quantity,
+                                       BigInteger quantity,
                                        BigDecimal price,
                                        BigDecimal sum,
                                        String itemsName,
