+++ src/main/java/ru/kolesnikov/storemanagement/controller/dto/shipmentdetails/ShipmentDetailDTOResponse.java
@@ -1,9 +1,10 @@
 package ru.kolesnikov.storemanagement.controller.dto.shipmentdetails;
 
 import java.math.BigDecimal;
+import java.math.BigInteger;
 
 public record ShipmentDetailDTOResponse(Long id,
-                                        BigDecimal quantity,
+                                        BigInteger quantity,
                                         BigDecimal price,
                                         BigDecimal sum,
                                         String itemsName,
